Add the ability for humans to heal their team members
The medkit handling is moved to a new function G_UseMedkit.
If there is a human within a given range that is either more wounded
than the player, or poisoned, the medkit is applied to him
(see G_NeedsMedkit).

The range and the breadth of the action is configurable by cvars.

The original patch was used on a server, but this port wasn't tested
(only that it compiles).